# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Makwa Modules's API",
        "contact": {
            "name": "sameh taha",
            "email": "smta0@yahoo.com"
        },
        "license": {
            "name": "Developed by sameh taha",
            "url": "https://www.makwaapp.com"
        },
        "version": "1.9.0"
    },
    "servers": [
        {
            "url": "{schema}://makwa.test",
            "description": "Development",
            "variables": {
                "schema": {
                    "enum": [
                        "https",
                        "http"
                    ],
                    "default": "http"
                }
            }
        },
        {
            "url": "{schema}://dev.makwaapp.com",
            "description": "Stage",
            "variables": {
                "schema": {
                    "enum": [
                        "https",
                        "http"
                    ],
                    "default": "https"
                }
            }
        },
        {
            "url": "{schema}://makwaapp.com",
            "description": "Production",
            "variables": {
                "schema": {
                    "enum": [
                        "https",
                        "http"
                    ],
                    "default": "https"
                }
            }
        }
    ],
    "paths": {
        "/index": {
            "get": {
                "summary": "list products",
                "operationId": "App\\Http\\Controllers\\ExampleController::index",
                "parameters": [
                    {
                        "name": "first_name",
                        "in": "query",
                        "description": "Tags to filter by customer name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "customer_id",
                        "in": "query",
                        "description": "Tags to filter by customer name",
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list with products"
                    },
                    "400": {
                        "description": "Invalid tag value"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/bpach": {
            "post": {
                "tags": [
                    "test"
                ],
                "summary": "list products",
                "operationId": "App\\Http\\Controllers\\ExampleController::post",
                "responses": {
                    "200": {
                        "description": "A list with products"
                    }
                }
            }
        },
        "/cpach": {
            "put": {
                "tags": [
                    "users"
                ],
                "summary": "add products",
                "operationId": "App\\Http\\Controllers\\ExampleController::put",
                "responses": {
                    "200": {
                        "description": "A list with products"
                    }
                }
            }
        },
        "/dindex": {
            "delete": {
                "tags": [
                    "users"
                ],
                "summary": "list products",
                "operationId": "App\\Http\\Controllers\\ExampleController::delete",
                "responses": {
                    "200": {
                        "description": "A list with products"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:pets",
                            "read:pets"
                        ]
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "petstore_auth": {
                "type": "oauth2",
                "flows": {
                    "implicit": {
                        "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
                        "scopes": {
                            "read:pets": "read your pets",
                            "write:pets": "modify pets in your account"
                        }
                    }
                }
            }
        }
    }
}